requirements.txt
streamlit>=1.38.0
tensorflow-cpu==2.19.0
numpy>=1.26.0
pickle-mixin
